+++ home.nix
@@ -12,6 +12,8 @@ in
     ./config/cli/htop.nix # htop monitor
     ./config/cli/cava.nix # Audio visualize Dracula theme (others in file)
     #./config/editors/nvf.nix # nvf alternate NVIM config 
+    ./config/noctalia.nix # Noctalia QuickShell wiring (fronm ddubsos)
+    ./config/overview.nix # Quickshell-overview workspace preview
     ./config/editors/vscode.nix # w/plugins and nero hyprland theme
     ./config/noctalia.nix # Noctalia QuickShell wiring (fronm ddubsos)
     ./config/terminals/kitty.nix #kitty term and kitty-bg (background in kitty)